mcpMode/mcpClient/requirements.txt
aiohttp>=3.8.0
aioconsole>=0.6.0
rich>=13.0.0
typing-extensions>=4.0.0
python-json-logger>=2.0.0
pyyaml>=6.0.1
